chore: refactor statistical analysis related endpoints
