+++ main.py
@@ -56,7 +56,7 @@ api_app.include_router(routes.general.router, prefix="")
 api_app.include_router(routes.table.router, prefix="/table/{project_id}")
 api_app.include_router(routes.topic.router, prefix="/topic/{project_id}")
 api_app.include_router(routes.userdata.router, prefix="/userdata/{project_id}")
-api_app.include_router(routes.statistic_test.router, prefix="/statistic-test/{project_id}")
+api_app.include_router(routes.statistical_analysis.router, prefix="/statistical-analysis/{project_id}")
 register_error_handlers(api_app)
 
 app.mount('/api', api_app)
