--- routes/statistic_test/controller/__init__.py
@@ -1,4 +1,0 @@
-from .pairwise import *
-from .basic import *
-from .crosstab import *
-from .regression import *
