+++ routes/statistical_analysis/controller/__init__.py
@@ -1,0 +1,4 @@
+from .binary_statistic_test import *
+from .statistic_test import *
+from .crosstab import *
+from .regression import *
