+++ routes/statistical_analysis/controller/statistic_test.py
@@ -1,12 +1,36 @@
 import itertools
-
-from modules.comparison.engine import TableComparisonEngine, StatisticTestResult
+from modules.comparison import TableComparisonEngine
+from modules.comparison.engine import StatisticTestResult
+from modules.config import SchemaColumnTypeEnum
 from modules.config.schema.base import ANALYZABLE_SCHEMA_COLUMN_TYPES
 from modules.logger.provisioner import ProvisionedLogger
 from modules.project.cache import ProjectCache
-from modules.table.filter_variants import NamedTableFilter, NotTableFilter
-from ..model import BinaryStatisticTestOnDistributionResultResource, BinaryStatisticTestOnDistributionSchema, PairwiseStatisticTestResultResource, PairwiseStatisticTestSchema
 from routes.table.controller.preprocess import TablePreprocessModule
+
+from ..model import (
+  PairwiseStatisticTestResultResource,
+  PairwiseStatisticTestSchema,
+  StatisticTestSchema,
+  OmnibusStatisticTestSchema,
+)
+
+def statistic_test(params: StatisticTestSchema, cache: ProjectCache):
+  config = cache.config
+  config.data_schema.assert_of_type(params.column, ANALYZABLE_SCHEMA_COLUMN_TYPES)
+  df = cache.workspaces.load()
+  engine = TableComparisonEngine(
+    config=config,
+    groups=[params.group1, params.group2],
+    exclude_overlapping_rows=params.exclude_overlapping_rows
+  )
+  result = engine.compare(
+    df,
+    column_name=params.column,
+    statistic_test_preference=params.statistic_test_preference,
+    effect_size_preference=params.effect_size_preference,
+  )
+  
+  return result
 
 def pairwise_statistic_test(cache: ProjectCache, input: PairwiseStatisticTestSchema):
   import scipy.stats
@@ -42,38 +66,30 @@ def pairwise_statistic_test(cache: ProjectCache, input: PairwiseStatisticTestSch
   )
 
 
-def binary_statistic_test_on_distribution(cache: ProjectCache, input: BinaryStatisticTestOnDistributionSchema):
-  preprocess = TablePreprocessModule(cache)
-  column = preprocess.assert_column(input.column, supported_types=ANALYZABLE_SCHEMA_COLUMN_TYPES)
-  workspace = cache.workspaces.load()
-
-  results: list[StatisticTestResult] = []
-  with ProvisionedLogger().disable(["TableEngine", "TableComparisonEngine"]):
-    for group in input.groups:
-      anti_group = NamedTableFilter(
-        name=f"NOT {group.name}",
-        filter=NotTableFilter(operand=group.filter)
-      )
-      engine = TableComparisonEngine(
-        config=cache.config,
-        groups=[group, anti_group],
-        exclude_overlapping_rows=False,
-      )
-      result = engine.compare(
-        workspace,
-        column_name=input.column,
-        statistic_test_preference=input.statistic_test_preference,
-        effect_size_preference=input.effect_size_preference,
-      )
-      results.append(result)
-  return BinaryStatisticTestOnDistributionResultResource(
-    column=column,
-    groups=list(map(lambda group: group.name, input.groups)),
-    results=results,
+def omnibus_statistic_test(cache: ProjectCache, input: OmnibusStatisticTestSchema):
+  config = cache.config
+  config.data_schema.assert_of_type(input.column, [
+    SchemaColumnTypeEnum.Categorical,
+    SchemaColumnTypeEnum.Continuous,
+    SchemaColumnTypeEnum.OrderedCategorical,
+    SchemaColumnTypeEnum.Temporal,
+    SchemaColumnTypeEnum.Topic,
+  ])
+  df = cache.workspaces.load()
+  engine = TableComparisonEngine(
+    config=config,
+    groups=input.groups,
+    exclude_overlapping_rows=input.exclude_overlapping_rows
   )
-
+  result = engine.compare_omnibus(
+    df,
+    column_name=input.column,
+    statistic_test_preference=input.statistic_test_preference,
+  )
+  return result
 
 __all__ = [
+  "statistic_test",
   "pairwise_statistic_test",
-  "binary_statistic_test_on_distribution"
+  "omnibus_statistic_test"
 ]
