+++ routes/statistical_analysis/model.py
